requirements.txt
pymongo
python-dotenv 